Dokončenie CRUDu

+++ php/crud/pages/users/list.php
@@ -1,4 +1,8 @@
-<a href="/index.php?p=users/add">Pridaj osobu</a>
+<?php
+$userStorage = new UserStorage();
+?>
+
+<a href="?p=users/add">Pridaj osobu</a>
 <br><br>
 
 <table border="1">
@@ -9,13 +13,13 @@
         <th>Krajina</th>
         <th>Akcie</th>
     </tr>
-    <?php foreach (UserStorage::getAllUsers() as $user) { ?>
+    <?php foreach ($userStorage->getAllUsers() as $user) { ?>
         <tr>
-            <td><?=$user->name?></td>
-            <td><?=$user->surname?></td>
-            <td><?=$user->mail?></td>
-            <td><?=$user->country?></td>
-            <td><a href="?p=users/edit&id=<?=$user->id?>">Edit</a> | <a href="?p=users/delete&id=<?=$user->id?>">Delete</a> </td>
+            <td><?=htmlentities($user->name)?></td>
+            <td><?=htmlentities($user->surname)?></td>
+            <td><?=htmlentities($user->mail)?></td>
+            <td><?=htmlentities($user->country)?></td>
+            <td><a href="?p=users/edit&id=<?=$user->id?>">Edit</a> | <a href="?p=users/delete&id=<?=$user->id?>" onclick="return confirm('Skutočne chcete odstrániť tento záznam?')">Delete</a> </td>
         </tr>
     <?php } ?>
 </table>
